app/pyproject.toml
[tool.poetry]
name = "reddit-sentiment-analysis"
version = "0.1.0"
description = "AI-Based Sentiment Analysis for Business Reviews on Reddit"
authors = ["Your Name <billnepill@gmail.com>"]
readme = "README.md"
packages = [{include = "reddit_sentiment_analysis", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = ">=0.3,<0.4"
langchain-openai = ">=0.3,<0.4"
langchain-core = ">=0.3,<0.4"
langgraph = ">=0.2.20,<0.3"
openai = "^1.1.1"
streamlit = "^1.29.0"
pandas = "^2.1.3"
praw = "^7.7.1"
python-dotenv = "^1.0.0"
faiss-cpu = "^1.7.4"
certifi = "^2025.1.31"  # For proper SSL certificate verification in email service
requests-cache = "^1.2.1"

[tool.poetry.scripts]
reddit-sentiment-analysis = "reddit_sentiment_analysis.gui:run_gui"
reddit-monitor = "reddit_sentiment_analysis.gui:run_gui"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
